Cleanup, refactoring
