+++ src/pages/annotate/Annotate.tsx
@@ -3,7 +3,8 @@ import { useNavigate, useParams } from 'react-router-dom';
 import { AnnotoriousManifold, OSDViewerManifold } from '@annotorious/react-manifold';
 import { Image, LoadedImage } from '@/model';
 import { useImages } from '@/store';
-import { HeaderSection, ToolMode, Tool } from './HeaderSection';
+import { HeaderSection } from './HeaderSection';
+import { ToolMode, Tool } from './Tool';
 import { SavingState } from './SavingState';
 import { SidebarSection } from './SidebarSection';
 import { WorkspaceSection} from './WorkspaceSection';
