+++ src/pages/annotate/Tool.ts
@@ -1,0 +1,3 @@
+export type Tool = 'rectangle' | 'polygon' | 'ellipse';
+
+export type ToolMode = 'move' | 'draw';
