+++ src/pages/annotate/WorkspaceSection/useDrawingStyles.ts
@@ -2,17 +2,17 @@ import type { DataModel } from '@/model';
 import { useDataModel } from '@/store';
 import { Color, DrawingStyle, ImageAnnotation, W3CAnnotationBody } from '@annotorious/react';
 
-export const colorByEntity = (
+export const colorByEntityType = (
   model: DataModel
 ) => (annotation: ImageAnnotation): DrawingStyle => {
 
-  const firstEntityBody: W3CAnnotationBody = 
+  const firstEntityBody: W3CAnnotationBody =
     annotation.bodies.find(b => b.purpose === 'classifying');
 
   if (firstEntityBody) {
-    const entity = model.entityTypes.find(e => e.id === firstEntityBody.source);
-    return entity ? { 
-      fill: entity.color as Color
+    const entityType = model.entityTypes.find(e => e.id === firstEntityBody.source);
+    return entityType ? { 
+      fill: entityType.color as Color
     } : {
       fill: '#000000'
     };
@@ -29,7 +29,7 @@ export const useDrawingStyles = () => {
   const { model } = useDataModel();
 
   return {
-    colorByEntity: colorByEntity(model)
+    colorByEntity: colorByEntityType(model)
   }
 
 }
